pyproject.toml
[tool.poetry]
name = "orders_list_app"
version = "0.1.0"
description = ""
authors = ["EvilMadSquirrel <minichev.s.l@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
python-dotenv = "^0.20.0"
google-api-python-client = "^2.51.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
requests = "^2.28.0"
xmltodict = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
